app/requirements.txt
# API Framework
fastapi>=0.100.0,<0.110.0
uvicorn>=0.21.0,<0.22.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0  # For BaseSettings

# Database
sqlalchemy>=2.0.25
greenlet>=3.0.3
alembic>=1.10.3,<1.11.0
psycopg2-binary>=2.9.6,<2.10.0

# Authentication
python-jose>=3.3.0,<3.4.0
passlib>=1.7.4,<1.8.0
bcrypt>=4.0.1,<4.1.0

# Agent Framework
camel-ai>=0.1.0,<0.2.0
Pillow>=10.0.0  # Required by camel-ai

# Browser Automation
playwright>=1.41.0

# Vector Database
chromadb>=0.3.21,<0.4.0

# Utilities
python-multipart>=0.0.6,<0.1.0
tenacity>=8.2.2,<8.3.0
python-dotenv>=1.0.0,<2.0.0 